// src/_common/database/database-repository.enum.ts
export enum Repositories {
  UsersRepository = 'UsersRepository',
  FcmTokensRepository = 'FcmTokensRepository',
  SessionsRepository = 'SessionsRepository',
  TweetRepository = 'TweetsRepository',
  FollowsRepository = 'FollowsRepository',
  UserVerificationCodesRepository = 'UserVerificationCodesRepository',
  SecurityGroupsRepository = 'SecurityGroupsRepository',
  AppConfigurationsRepository = 'AppConfigurationsRepository',
  FilesRepository = 'FilesRepository',
}